extract 서버 없이도 작동하게 하여 실행 속도 최적화
